pyproject.toml
[tool.poetry]
name = "proxy-service"
version = "0.1.0"
description = ""
authors = ["Pavel Nikulenok <pashaaaa86@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.32.3"
prometheus-client = "^0.21.1"
python-dotenv = "^1.0.1"
httpx = "^0.28.1"
uvicorn = "^0.34.0"
drfasyncview = "^0.2.3"
django-extensions = "^3.2.3"
redis = "^5.2.1"
django-redis = "^5.4.0"
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pytest-asyncio = "^0.25.0"
django-prometheus = "^2.3.1"
sentry-sdk = {extras = ["django"], version = "^2.19.2"}

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
